Extract the component that needs to client-side. The rest can SSR.
